gen/requirement.txt
gradio
openai
requests
pillow
python-dotenv
flask
stripe
snowflake-id